--- e2e_tests/tests/fixtures/pytorch_identity/distributed.yaml
@@ -1,24 +1,0 @@
-name: pytorch-identity
-data:
-  model_type: single_output
-entrypoint: model_def:IdentityPyTorchTrial
-hyperparameters:
-  global_batch_size: 4
-records_per_epoch: 64
-resources:
-  slots_per_trial: 2
-scheduling_unit: 1
-searcher:
-  name: single
-  metric: val_loss
-  max_length:
-    epochs: 3
-  smaller_is_better: true
-max_restarts: 0
-min_checkpoint_period:
-  batches: 1
-min_validation_period:
-  batches: 1
-optimizations:
-  aggregation_frequency: 2
-  average_aggregated_gradients: true
