--- harness/tests/experiment/fixtures/pytorch_identity/distributed.yaml
@@ -1,1 +1,0 @@
-../../../../../e2e_tests/tests/fixtures/pytorch_identity/distributed.yaml
